Add files via upload
The first is self explanatory. It has the paths to copy into the app.yaml
The second is for the most part complete, with the exception the $scope.user_profile_id and the $scope.baseUrl need to be changed accordingly. I will upload this file again at midnight with the cookie code commented out.